requirements/dev.txt
django-nose==1.0
nose==1.1.2
django-coverage==1.2.2
django-debug-toolbar==0.9.4
Paver==1.0.5
